+++ tests/test_suite.py
@@ -5,6 +5,7 @@ import os
 import json
 import re
 import pytest
+import time
 
 dir_name = os.path.dirname(__file__)
 main = os.path.join(dir_name, '..')
@@ -20,6 +21,29 @@ sys.path.append(services)
 
 from a4kSubtitles import api
 from tests import utils
+
+__movie_video_meta = {
+    'year': '2016',
+    'title': 'Fantastic Beasts and Where to Find Them',
+    'imdb_id': 'tt3183660',
+    'filename': 'Fantastic.Beasts.and.Where.to.Find.Them.2016.1080p.BluRay.x264.DTS-JYK.mkv',
+    'filesize': '3592482379',
+    'filehash': '4985126cbf92fe60',
+    'subdb_hash': 'fb7e5d6ac9c3f94813467988de753d0e',
+}
+
+__tvshow_video_meta = {
+    "year": "2018",
+    "title": "The Passenger",
+    "tvshow": "Westworld",
+    "imdb_id": "tt6243312",
+    "season": "2",
+    "episode": "10",
+    "filename": "Westworld.S02E10.1080p.WEB.H264-DEFLATE.mkv",
+    "filesize": "7945997565",
+    "filehash": "d603a5b0e73d4b6b",
+    "subdb_hash": "2aec1b70afe702e67ab39a0af776ba5a",
+}
 
 def __remove_last_results(a4ksubtitles_api):
     try:
@@ -46,20 +70,22 @@ def __search(a4ksubtitles_api, settings={}, video_meta={}):
     }
     search.settings.update(settings)
 
-    search.video_meta = {
-        'year': '2016',
-        'title': 'Fantastic Beasts and Where to Find Them',
-        'imdb_id': 'tt3183660',
-        'filename': 'Fantastic.Beasts.and.Where.to.Find.Them.2016.1080p.BluRay.x264.DTS-JYK.mkv',
-        'filesize': '3592482379',
-        'filehash': '4985126cbf92fe60',
-        'subdb_hash': 'fb7e5d6ac9c3f94813467988de753d0e'
-    }
+    search.video_meta = {}
     search.video_meta.update(video_meta)
 
     search.results = a4ksubtitles_api.search(search.params, search.settings, search.video_meta)
 
     return search
+
+def __search_movie(a4ksubtitles_api, settings={}, video_meta={}):
+    movie_video_meta = __movie_video_meta.copy()
+    movie_video_meta.update(video_meta)
+    return __search(a4ksubtitles_api, settings, movie_video_meta)
+
+def __search_tvshow(a4ksubtitles_api, settings={}, video_meta={}):
+    tvshow_video_meta = __tvshow_video_meta.copy()
+    tvshow_video_meta.update(video_meta)
+    return __search(a4ksubtitles_api, settings, tvshow_video_meta)
 
 def test_api():
     def get_error_msg(e):
@@ -110,7 +136,7 @@ def test_opensubtitles():
         'opensubtitles.username': os.getenv('A4KSUBTITLES_OPENSUBTITLES_USERNAME', ''),
         'opensubtitles.password': os.getenv('A4KSUBTITLES_OPENSUBTITLES_PASSWORD', '')
     }
-    search = __search(a4ksubtitles_api, settings)
+    search = __search_movie(a4ksubtitles_api, settings)
 
     assert len(search.results) == 20
 
@@ -125,7 +151,7 @@ def test_opensubtitles():
         'filesize': '',
         'filehash': '',
     }
-    search = __search(a4ksubtitles_api, settings, video_meta)
+    search = __search_movie(a4ksubtitles_api, settings, video_meta)
 
     assert len(search.results) == 20
 
@@ -159,6 +185,29 @@ def test_opensubtitles():
 
     assert re.match(r'.*OpenSubtitles.*', sub_contents, re.DOTALL) is None
 
+def test_opensubtitles_tvshow():
+    a4ksubtitles_api = api.A4kSubtitlesApi({'kodi': True})
+    __remove_last_results(a4ksubtitles_api)
+
+    # search
+    settings = {
+        'opensubtitles.enabled': 'true',
+    }
+    search = __search_tvshow(a4ksubtitles_api, settings)
+
+    # download
+    item = search.results[0]
+
+    params = {
+        'action': 'download',
+        'service_name': 'opensubtitles',
+        'action_args': item['action_args']
+    }
+
+    filepath = a4ksubtitles_api.download(params, search.settings)
+
+    assert filepath != ''
+
 def test_bsplayer():
     a4ksubtitles_api = api.A4kSubtitlesApi({'kodi': True})
     __remove_last_results(a4ksubtitles_api)
@@ -167,7 +216,7 @@ def test_bsplayer():
     settings = {
         'bsplayer.enabled': 'true',
     }
-    search = __search(a4ksubtitles_api, settings)
+    search = __search_movie(a4ksubtitles_api, settings)
 
     assert len(search.results) == 16
 
@@ -177,11 +226,34 @@ def test_bsplayer():
 
     # cache
     request_execute_spy = utils.spy_fn(a4ksubtitles_api.core.request, 'execute')
-    __search(a4ksubtitles_api, settings)
+    __search_movie(a4ksubtitles_api, settings)
 
     assert request_execute_spy.call_count == 0
 
     request_execute_spy.restore()
+
+    # download
+    item = search.results[0]
+
+    params = {
+        'action': 'download',
+        'service_name': 'bsplayer',
+        'action_args': item['action_args']
+    }
+
+    filepath = a4ksubtitles_api.download(params, search.settings)
+
+    assert filepath != ''
+
+def test_bsplayer_tvshow():
+    a4ksubtitles_api = api.A4kSubtitlesApi({'kodi': True})
+    __remove_last_results(a4ksubtitles_api)
+
+    # search
+    settings = {
+        'bsplayer.enabled': 'true',
+    }
+    search = __search_tvshow(a4ksubtitles_api, settings)
 
     # download
     item = search.results[0]
@@ -204,7 +276,30 @@ def test_podnadpisi():
     settings = {
         'podnadpisi.enabled': 'true',
     }
-    search = __search(a4ksubtitles_api, settings)
+    search = __search_movie(a4ksubtitles_api, settings)
+
+    # download
+    item = search.results[0]
+
+    params = {
+        'action': 'download',
+        'service_name': 'podnadpisi',
+        'action_args': item['action_args']
+    }
+
+    filepath = a4ksubtitles_api.download(params, search.settings)
+
+    assert filepath != ''
+
+def test_podnadpisi_tvshow():
+    a4ksubtitles_api = api.A4kSubtitlesApi({'kodi': True})
+    __remove_last_results(a4ksubtitles_api)
+
+    # search
+    settings = {
+        'podnadpisi.enabled': 'true',
+    }
+    search = __search_tvshow(a4ksubtitles_api, settings)
 
     # download
     item = search.results[0]
@@ -227,7 +322,31 @@ def test_subdb():
     settings = {
         'subdb.enabled': 'true',
     }
-    search = __search(a4ksubtitles_api, settings)
+    search = __search_movie(a4ksubtitles_api, settings)
+
+    # download
+    item = search.results[0]
+
+    params = {
+        'action': 'download',
+        'service_name': 'subdb',
+        'action_args': item['action_args']
+    }
+
+    filepath = a4ksubtitles_api.download(params, search.settings)
+
+    assert filepath != ''
+
+def test_subdb_tvshow():
+    a4ksubtitles_api = api.A4kSubtitlesApi({'kodi': True})
+    __remove_last_results(a4ksubtitles_api)
+
+    # search
+    settings = {
+        'subdb.enabled': 'true',
+    }
+
+    search = __search_tvshow(a4ksubtitles_api, settings)
 
     # download
     item = search.results[0]
@@ -250,7 +369,7 @@ def test_subscene():
     settings = {
         'subscene.enabled': 'true',
     }
-    search = __search(a4ksubtitles_api, settings)
+    search = __search_movie(a4ksubtitles_api, settings)
 
     # download
     item = search.results[0]
@@ -260,6 +379,39 @@ def test_subscene():
         'service_name': 'subscene',
         'action_args': item['action_args']
     }
+
+    if os.getenv('CI', None) is not None:
+        time.sleep(2)
+
+    filepath = a4ksubtitles_api.download(params, search.settings)
+
+    assert filepath != ''
+
+def test_subscene_tvshow():
+    a4ksubtitles_api = api.A4kSubtitlesApi({'kodi': True})
+    __remove_last_results(a4ksubtitles_api)
+
+    # search
+    settings = {
+        'subscene.enabled': 'true',
+    }
+
+    if os.getenv('CI', None) is not None:
+        time.sleep(2)
+
+    search = __search_tvshow(a4ksubtitles_api, settings)
+
+    # download
+    item = search.results[0]
+
+    params = {
+        'action': 'download',
+        'service_name': 'subscene',
+        'action_args': item['action_args']
+    }
+
+    if os.getenv('CI', None) is not None:
+        time.sleep(2)
 
     filepath = a4ksubtitles_api.download(params, search.settings)